DataSimulator/requirements.txt
alabaster==0.7.7
anaconda-client==1.4.0
anaconda-navigator==1.1.0
appdirs==1.4.3
appnope==0.1.0
appscript==1.0.1
argcomplete==1.0.0
astropy==1.1.2
Babel==2.2.0
backports-abc==0.4
backports.ssl-match-hostname==3.4.0.2
beautifulsoup4==4.4.1
bitarray==0.8.1
blaze==0.9.1
bokeh==0.11.1
boto==2.39.0
Bottleneck==1.0.0
cachetools==2.0.0
cdecimal==2.3
certifi==2017.4.17
cffi==1.5.2
chardet==3.0.3
chest==0.2.3
cloudpickle==0.1.1
clyent==1.2.1
colorama==0.3.7
conda==4.2.13
conda-build==0+unknown
conda-manager==0.3.1
configobj==5.0.6
cryptography==1.7.2
cycler==0.10.0
Cython==0.23.4
cytoolz==0.7.5
dask==0.8.1
datashape==0.5.1
decorator==4.0.9
dill==0.2.6
docutils==0.12
dynd==0.7.3.dev1
enum34==1.1.2
et-xmlfile==1.0.1
fastcache==1.0.2
Flask==0.10.1
Flask-Cors==2.1.2
funcsigs==0.4
future==0.16.0
futures==3.0.3
gapic-google-cloud-pubsub-v1==0.15.4
gevent==1.1.0
google-api-python-client==1.6.2
google-auth==1.0.1
google-auth-httplib2==0.0.2
google-cloud-core==0.24.1
google-cloud-pubsub==0.25.0
google-gax==0.15.13
googleapis-common-protos==1.5.2
greenlet==0.4.9
grin==1.2.1
grpc-google-iam-v1==0.11.1
grpcio==1.3.5
h5py==2.5.0
HeapDict==1.0.0
httplib2==0.10.3
idna==2.5
ipaddress==1.0.14
ipykernel==4.3.1
ipython==4.1.2
ipython-genutils==0.1.0
ipywidgets==4.1.1
itsdangerous==0.24
jdcal==1.2
jedi==0.9.0
Jinja2==2.8
jsonschema==2.4.0
jupyter==1.0.0
jupyter-client==4.2.2
jupyter-console==4.1.1
jupyter-core==4.1.0
jwt==0.5.2
llvmlite==0.9.0
locket==0.2.0
lxml==3.6.0
MarkupSafe==0.23
matplotlib==1.5.1
mistune==0.7.2
mpmath==0.19
multipledispatch==0.4.8
nbconvert==4.1.0
nbformat==4.0.1
networkx==1.11
nltk==3.2
nose==1.3.7
notebook==4.1.0
numba==0.24.0
numexpr==2.5
numpy==1.10.4
oauth2client==3.0.0
odo==0.4.2
openpyxl==2.3.2
packaging==16.8
paho-mqtt==1.2.3
pandas==0.18.0
partd==0.3.2
path.py==0.0.0
patsy==0.4.0
pep8==1.7.0
pexpect==4.0.1
pickleshare==0.5
Pillow==3.1.1
ply==3.8
proto-google-cloud-pubsub-v1==0.15.4
protobuf==3.3.0
psutil==4.1.0
ptyprocess==0.5
py==1.4.31
pyasn1==0.1.9
pyasn1-modules==0.0.8
PyAudio==0.2.7
pycosat==0.6.1
pycparser==2.14
pycrypto==2.6.1
pycurl==7.19.5.1
pyflakes==1.1.0
Pygments==2.1.1
PyJWT==1.5.0
pyOpenSSL==0.15.1
pyparsing==2.2.0
pytest==2.8.5
python-dateutil==2.5.1
pytz==2016.2
PyYAML==3.11
pyzmq==15.2.0
QtAwesome==0.3.2
qtconsole==4.2.0
QtPy==1.0
redis==2.10.3
requests==2.17.3
rope==0.9.4
rsa==3.4.2
scikit-image==0.12.3
scikit-learn==0.17.1
scipy==0.17.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
snowballstemmer==1.2.1
sockjs-tornado==1.0.1
Sphinx==1.3.5
sphinx-rtd-theme==0.1.9
spyder==2.3.8
SQLAlchemy==1.0.12
statsmodels==0.6.1
sympy==1.0
tables==3.2.2
terminado==0.5
toolz==0.7.4
tornado==4.3
traitlets==4.2.1
typing==3.5.3.0
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.21.1
virtualenv==15.1.0
Werkzeug==0.11.4
xlrd==0.9.4
XlsxWriter==0.8.4
xlwings==0.7.0
xlwt==1.0.0
